+++ packages/atomic-react/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.9](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@2.2.8...@coveo/atomic-react@2.2.9) (2023-04-12)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+## [2.2.8](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@2.2.7...@coveo/atomic-react@2.2.8) (2023-04-12)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+## [2.2.7](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@2.2.6...@coveo/atomic-react@2.2.7) (2023-04-12)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
 ## [2.2.6](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@2.2.5...@coveo/atomic-react@2.2.6) (2023-04-11)
 
 **Note:** Version bump only for package @coveo/atomic-react